react-django/requirements.txt
asn1crypto==1.0.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.2.0
autopep8==1.4.4
cffi==1.12.3
colorama==0.4.1
cryptography==2.7
Django==2.2.6
django-rest-knox==4.1.0
django-rest-passwordreset==1.0.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.2
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
npm==0.1.1
optional-django==0.1.0
packaging==19.2
Pillow==6.2.0
pluggy==0.13.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pyparsing==2.4.2
pytest==5.2.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0